WE GOT A WORKING CABLE PROTOTPYE FELLAS
cable is working great! Added a ton of exports so that it can be used in the editor and seen live, and so that end points can be changed, as well as the sprite or line used for the cable. Still more to come once we decide to integrate it.....