Final Physics edits to the cable for now
Did more testing with rigid body chains using the pinjoints and even then the pin joints get easily seperated by static bodies. I did a ton of different combos of rigid, static, and kinematic bodies and this seemes to be the best approach. Getting to rope to work properly will be a matter of adjusting values and the length and width of the collision shape.