Added Plug Concept and Capability to Cable
Added the ability for cables to have plug ends, and made a plug be a resource. The idea will be that there will be different scene packages for cables with already designated plugs, textures, nodes counts, lengths, etc.

Need to still add capability to connect to a plugs and ports and keep track of what is static and what pins terminations are static on the cable