Refactored an important plug method
Not in the dark anymore about which plug goes first and what not. All is accounted for and easier to read.

Also added that any plug being grabbed by astro uses astro's item area to look for other plugs when interacting again

Still need to figure out what to do when more than three plugs are available in the same space. Also need to make combining cables a thing