ton of shit - refactored interactNodes - global bug fix - plug text works
1) Refactored interactNodes:
I realized that a better way to manage multiple dynamic text lines would be to build a system around the interactNode (which asts as a reference for scene heiarchy and instancing) to have multiple interactNodes, including a special designated infoInteractNode that is meant for displaying info based on the other interact nodes. Had to build several systems for calculating where to place interactNodes, which to use, destroy, which way to face and force facing depending on what is already drawn, etc. Was a bitch, especially with calculating the correct amount of lines a text created -- I had to make my own method cause the two built-in methods I could have used didn't work properly (confirmed with godot community). One you had use call_deferred or else it would give you the memory address, and the other would give you the numeber of lines, but it wouldn't calculate wrapping logic properly.

2) Global bug fix
In the tween and timer managers I have, they return the tween or timer for the node using them to manage if they are to be destroyed prematurely. A long time ago I had run into a bug where the system could occasionally be trying to prematurely remove something that wasn't there in memory, and consequently pointing to a wrong and different node, timer, or tween. I tried to make up for this by checking if it was of the expected type (tween or timer) and if it was the same node as the one I had pointed to originally, but I was running into a weird bug and realized I could check its original memory name by checking against the original node.to_string() which includes its unique memory id. That solved my problem, and I went back and added this id check to where ever nodes were prematurely destroying their tween or timer.

3) Plug text works:
Plug text fucking works! To display the right text (after modifying the interactNodes system, I had to add a "hypothetical" bool to checking results, in which the hypothetical would give me the actual result if I were to hit interact, but without actually changing anything. But because this hypothetical check was not the same as hitting interact on all plug nodes at once and marking some as processed, the way I followed the logic after implementing the hypothetical was by looking for what I wanted, making note of the conditions it spit out, and then reverse engineering that into the hypothetical logic (with the addition of two new enum values), and then checked that against a giant switch statement (match statement, in godot). Also made the cable texts by default read correctly depending on the plugs attributes.

I spent the whole damn day doing all of this shit and I think it looks siiiiick. Just need to fix some minor physics bugs where the cable jumps when you connect them (probably because of reversal) but I can do that in the clean up for sure. Cables are OFFICIALLY DONE MOTHER FUCKERRR