cable characterswitching restoring state works
got restoring states for cables and plugs to finally work properly, what a bitch with all the things to keep track of.

also fixed a bug where I the file path for loading lvl from a packed scene was null, so had to go back to old way where lvl node path is tracked in astro character res.